Separate packages for clj and cljs
There's enough code for each plugin that it's worth just having a
consistent separation.
