+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/clojure/tasks/ClojureForkOptions.java
@@ -1,0 +1,6 @@
+package gradle_clojure.plugin.clojure.tasks;
+
+import org.gradle.api.tasks.compile.BaseForkOptions;
+
+public class ClojureForkOptions extends BaseForkOptions {
+}
