+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/clojurescript/ClojurescriptPlugin.java
@@ -1,0 +1,13 @@
+package gradle_clojure.plugin.clojurescript;
+
+import org.gradle.api.Plugin;
+import org.gradle.api.Project;
+import org.gradle.api.plugins.JavaPlugin;
+
+public class ClojurescriptPlugin implements Plugin<Project> {
+  @Override
+  public void apply(Project project) {
+    project.getPlugins().apply(ClojurescriptBasePlugin.class);
+    project.getPlugins().apply(JavaPlugin.class);
+  }
+}
