+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/clojurescript/tasks/CheckedArrays.java
@@ -1,0 +1,5 @@
+package gradle_clojure.plugin.clojurescript.tasks;
+
+public enum CheckedArrays {
+  warn, error
+}
