+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/clojurescript/tasks/ClojurescriptSourceSet.java
@@ -1,0 +1,10 @@
+package gradle_clojure.plugin.clojurescript.tasks;
+
+import org.gradle.api.Action;
+import org.gradle.api.file.SourceDirectorySet;
+
+public interface ClojurescriptSourceSet {
+  SourceDirectorySet getClojurescript();
+
+  ClojurescriptSourceSet clojurescript(Action<? super SourceDirectorySet> configureAction);
+}
