+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/clojurescript/tasks/ModuleType.java
@@ -1,0 +1,5 @@
+package gradle_clojure.plugin.clojurescript.tasks;
+
+public enum ModuleType {
+  commonjs, amd, es6
+}
