+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/clojurescript/tasks/Optimizations.java
@@ -1,0 +1,5 @@
+package gradle_clojure.plugin.clojurescript.tasks;
+
+public enum Optimizations {
+  none, whitespace, simple, advanced
+}
