+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/common/internal/ExperimentalSettings.java
@@ -1,0 +1,11 @@
+package gradle_clojure.plugin.common.internal;
+
+public final class ExperimentalSettings {
+  private ExperimentalSettings() {
+    // do not instantiate
+  }
+
+  public static boolean isUseWorkers() {
+    return Boolean.getBoolean("gradle-clojure.experimental.use-workers");
+  }
+}
