--- modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/tasks/ClojureForkOptions.java
@@ -1,6 +1,0 @@
-package gradle_clojure.plugin.tasks;
-
-import org.gradle.api.tasks.compile.BaseForkOptions;
-
-public class ClojureForkOptions extends BaseForkOptions {
-}
