--- modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/tasks/ReflectionWarnings.java
@@ -1,44 +1,0 @@
-package gradle_clojure.plugin.tasks;
-
-import java.io.Serializable;
-
-import org.gradle.api.tasks.Input;
-
-public final class ReflectionWarnings implements Serializable {
-  private boolean enabled;
-  private boolean projectOnly;
-  private boolean asErrors;
-
-  public ReflectionWarnings(boolean enabled, boolean projectOnly, boolean asErrors) {
-    this.enabled = enabled;
-    this.projectOnly = projectOnly;
-    this.asErrors = asErrors;
-  }
-
-  @Input
-  public boolean isEnabled() {
-    return enabled;
-  }
-
-  public void setEnabled(boolean enabled) {
-    this.enabled = enabled;
-  }
-
-  @Input
-  public boolean isProjectOnly() {
-    return projectOnly;
-  }
-
-  public void setProjectOnly(boolean projectOnly) {
-    this.projectOnly = projectOnly;
-  }
-
-  @Input
-  public boolean isAsErrors() {
-    return asErrors;
-  }
-
-  public void setAsErrors(boolean asErrors) {
-    this.asErrors = asErrors;
-  }
-}
