--- modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/tasks/clojurescript/CheckedArrays.java
@@ -1,5 +1,0 @@
-package gradle_clojure.plugin.tasks.clojurescript;
-
-public enum CheckedArrays {
-  warn, error
-}
