--- modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/tasks/clojurescript/Optimizations.java
@@ -1,5 +1,0 @@
-package gradle_clojure.plugin.tasks.clojurescript;
-
-public enum Optimizations {
-  none, whitespace, simple, advanced
-}
